fix: cleaned up, added createdAt index
