+++ src/core/server/models/comment/index.ts
@@ -21,10 +21,11 @@ import {
   OrderedConnectionInput,
   resolveConnection,
 } from "talk-server/models/helpers/connection";
-import Query, {
+import {
   createConnectionOrderVariants,
   createIndexFactory,
-} from "talk-server/models/helpers/query";
+} from "talk-server/models/helpers/indexing";
+import Query from "talk-server/models/helpers/query";
 import { TenantResource } from "talk-server/models/tenant";
 import { CommentTag } from "./tag";
